requirements.txt
#
requests
APScheduler
fastapi
uvicorn
spotipy
python-dotenv